+++ ha-raumkernel-addon/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2.9
+- Automatic install of integration 
+
+## 1.2.7
+- Add Seek 
+- Improved Zone Handling 
+- Reboot Devices 
+
 ## 1.2.2
 
 - Add reboot feature to restart Raumfeld devices via SSH